+++ BitmovinAnalyticsCollector/Classes/Collector/data/EventData.swift
@@ -61,6 +61,8 @@ public class EventData: Codable {
     var pageLoadTime: Int64 = 0
     var version: String?
     var sequenceNumber: Int32 = 0
+    var drmType: String?
+    var drmLoadTime: Int64?
     #if os(iOS)
     var platform: String = "iOS"
     #elseif os(tvOS)